Bugfix: cat nc chunks in chrono order
Forgot that DataManager.data_files was a MultiMap, so files associated
with each DataKey were being stored in sets. Arguments to ncrcat in
fetch_dataset in gfdl.py have to be in the order the files should be
joined, so the bug was fixed by listing the files with the desired order
in fetch_dataset.