Add env var switch for 3D vs. 4D data
NOTE this is a temporary fix, identified in code with "HACK". In Python
driver scripts for MJO_suite and MJO_teleconnection, set env vars
depending on whether we have 3D or 4D data (based on string matching in
filenames only) , and use that to switch between pre-existing code in
NCL files.