Preallocate arrays as numpy.MaskedArrays
In the various get_* functions in COMPOSITE, averages are constructed by
taking the elementwise ratio of "dataout" and "ss" arrays. read_netcdf*
returns a MaskedArray which is generated by calling copy() on the input
array, so the input array must be a MaskedArray. To ensure this, replace
all instances where arrays are preallocated as np.zeros with
np.ma.zeros.

In COMPOSITE/get_flux_in_24.py, add missing statement to flag invalid
values in vvar as those greater than the sentinel value.